move cert related code to its own file
No code changes.
