+++ pkg/utils/auth.go
@@ -17,19 +17,19 @@ limitations under the License.
 package utils
 
 import (
-	"bytes"
 	"crypto/aes"
 	"crypto/cipher"
-	"crypto/rand"
-	"crypto/rsa"
-	"crypto/tls"
-	"crypto/x509"
-	"crypto/x509/pkix"
 	"encoding/base64"
 	"encoding/json"
-	"encoding/pem"
 	"errors"
 	"fmt"
+	"io/ioutil"
+	"os"
+	"path/filepath"
+	"regexp"
+	"strings"
+	"time"
+
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk"
 	"github.com/aliyun/alibaba-cloud-sdk-go/sdk/auth"
 	cre "github.com/aliyun/alibaba-cloud-sdk-go/sdk/auth/credentials"
@@ -37,14 +37,6 @@ import (
 	"github.com/emirpasic/gods/sets/hashset"
 	oidc "github.com/kubernetes-sigs/alibaba-cloud-csi-driver/pkg/auth"
 	log "github.com/sirupsen/logrus"
-	"google.golang.org/grpc/credentials"
-	"io/ioutil"
-	"math/big"
-	"os"
-	"path/filepath"
-	"regexp"
-	"strings"
-	"time"
 )
 
 const (
@@ -87,22 +79,6 @@ type ManageTokens struct {
 	RoleAccessKeySecret string
 	// RoleArn key
 	RoleArn string
-}
-
-// KeyPairArtifacts is cert struct
-type KeyPairArtifacts struct {
-	Cert    *x509.Certificate
-	Key     *rsa.PrivateKey
-	CertPEM []byte
-	KeyPEM  []byte
-}
-
-// CertOption is cert option
-type CertOption struct {
-	CAName          string
-	CAOrganizations []string
-	DNSNames        []string
-	CommonName      string
 }
 
 // AccessControlMode is int, represents different modes
@@ -480,159 +456,6 @@ func GetDefaultRoleAK() AccessControl {
 		return AccessControl{AccessKeyID: tokens.RoleAccessKeyID, AccessKeySecret: tokens.RoleAccessKeySecret, RoleArn: tokens.RoleArn, UseMode: ManagedToken}
 	}
 	return AccessControl{AccessKeyID: roleAccessKeyID, AccessKeySecret: roleAccessKeySecret, RoleArn: roleArn, UseMode: RoleArnToken}
-}
-
-// CreateCACert function is create cacert
-func CreateCACert(option CertOption, begin, end time.Time) (*KeyPairArtifacts, error) {
-	templ := &x509.Certificate{
-		SerialNumber: big.NewInt(0),
-		Subject: pkix.Name{
-			CommonName:   option.CAName,
-			Organization: option.CAOrganizations,
-		},
-		DNSNames:              option.DNSNames,
-		NotBefore:             begin,
-		NotAfter:              end,
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment | x509.KeyUsageCertSign,
-		BasicConstraintsValid: true,
-		IsCA:                  true,
-	}
-	key, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		return nil, fmt.Errorf("generating key: %s", err)
-	}
-	der, err := x509.CreateCertificate(rand.Reader, templ, templ, key.Public(), key)
-	if err != nil {
-		return nil, fmt.Errorf("creating certificate: %s", err)
-	}
-	certPEM, keyPEM, err := pemEncode(der, key)
-	if err != nil {
-		return nil, fmt.Errorf("encoding PEM: %s", err)
-	}
-	cert, err := x509.ParseCertificate(der)
-	if err != nil {
-		return nil, fmt.Errorf("parsing certificate: %s", err)
-	}
-
-	return &KeyPairArtifacts{Cert: cert, Key: key, CertPEM: certPEM, KeyPEM: keyPEM}, nil
-}
-
-// CreateCertPEM function is create cacert pem
-func CreateCertPEM(option CertOption, ca *KeyPairArtifacts, begin, end time.Time, isClient bool) ([]byte, []byte, error) {
-	sn, err := genSerialNum()
-	if err != nil {
-		return nil, nil, err
-	}
-	eks := []x509.ExtKeyUsage{x509.ExtKeyUsageServerAuth}
-	dnsNames := option.DNSNames
-	if isClient {
-		eks = []x509.ExtKeyUsage{x509.ExtKeyUsageClientAuth}
-		dnsNames = nil
-	}
-	templ := &x509.Certificate{
-		SerialNumber: sn,
-		Subject: pkix.Name{
-			CommonName: option.CommonName,
-		},
-		DNSNames:              dnsNames,
-		NotBefore:             begin,
-		NotAfter:              end,
-		KeyUsage:              x509.KeyUsageDigitalSignature | x509.KeyUsageKeyEncipherment,
-		ExtKeyUsage:           eks,
-		BasicConstraintsValid: true,
-	}
-	key, err := rsa.GenerateKey(rand.Reader, 2048)
-	if err != nil {
-		return nil, nil, fmt.Errorf("generating key: %s", err)
-	}
-	der, err := x509.CreateCertificate(rand.Reader, templ, ca.Cert, key.Public(), ca.Key)
-	if err != nil {
-		return nil, nil, fmt.Errorf("creating certificate: %s", err)
-	}
-	certPEM, keyPEM, err := pemEncode(der, key)
-	if err != nil {
-		return nil, nil, fmt.Errorf("encoding PEM: %s", err)
-	}
-	return certPEM, keyPEM, nil
-}
-
-func pemEncode(certificateDER []byte, key *rsa.PrivateKey) ([]byte, []byte, error) {
-	certBuf := &bytes.Buffer{}
-	if err := pem.Encode(certBuf, &pem.Block{Type: "CERTIFICATE", Bytes: certificateDER}); err != nil {
-		return nil, nil, fmt.Errorf("encoding cert: %s", err)
-	}
-	keyBuf := &bytes.Buffer{}
-	if err := pem.Encode(keyBuf, &pem.Block{Type: "RSA PRIVATE KEY", Bytes: x509.MarshalPKCS1PrivateKey(key)}); err != nil {
-		return nil, nil, fmt.Errorf("encoding key: %s", err)
-	}
-	return certBuf.Bytes(), keyBuf.Bytes(), nil
-}
-
-func genSerialNum() (*big.Int, error) {
-	serialNumLimit := new(big.Int).Lsh(big.NewInt(1), 128)
-	serialNum, err := rand.Int(rand.Reader, serialNumLimit)
-	if err != nil {
-		return nil, fmt.Errorf("serial number generation failure (%v)", err)
-	}
-	return serialNum, nil
-}
-
-// NewClientTLSFromFile function is new client with tls
-func NewClientTLSFromFile(serverName, caFile, certFile, keyFile string) (credentials.TransportCredentials, error) {
-	// Load the certificates from disk
-	certificate, err := tls.LoadX509KeyPair(certFile, keyFile)
-	if err != nil {
-		return nil, fmt.Errorf("could not load server key pair: %s", err)
-	}
-
-	// Create a certificate pool from the certificate authority
-	certPool := x509.NewCertPool()
-	ca, err := ioutil.ReadFile(caFile)
-	if err != nil {
-		return nil, fmt.Errorf("could not read ca certificate: %s", err)
-	}
-
-	// Append the client certificates from the CA
-	if ok := certPool.AppendCertsFromPEM(ca); !ok {
-		return nil, errors.New("failed to append client certs")
-	}
-
-	// Create the TLS credentials
-	creds := credentials.NewTLS(&tls.Config{
-		ServerName:   serverName,
-		Certificates: []tls.Certificate{certificate},
-		RootCAs:      certPool,
-	})
-	return creds, nil
-}
-
-// NewServerTLSFromFile function is new server with tls
-func NewServerTLSFromFile(caFile, certFile, keyFile string) (credentials.TransportCredentials, error) {
-	// Load the certificates from disk
-	certificate, err := tls.LoadX509KeyPair(certFile, keyFile)
-	if err != nil {
-		return nil, fmt.Errorf("could not load server key pair: %s", err)
-	}
-
-	// Create a certificate pool from the certificate authority
-	certPool := x509.NewCertPool()
-	ca, err := ioutil.ReadFile(caFile)
-	if err != nil {
-		return nil, fmt.Errorf("could not read ca certificate: %s", err)
-	}
-
-	// Append the client certificates from the CA
-	if ok := certPool.AppendCertsFromPEM(ca); !ok {
-		return nil, errors.New("failed to append client certs")
-	}
-
-	// Create the TLS credentials
-	creds := credentials.NewTLS(&tls.Config{
-		ClientAuth:   tls.RequireAndVerifyClientCert,
-		Certificates: []tls.Certificate{certificate},
-		ClientCAs:    certPool,
-	})
-	return creds, nil
 }
 
 // getCredentialAK get credential and config from credential files.
